feat: implement transfer source extents
When Polygon elements with linear reaction types deliver reactions through a rank 0 polygon, we need to know where on the source geometry the resulting distributed load is coming from. This allows use to ensure that the portion of the load within that region of the polygon is assigned correctly to the element it is transferring to.